#17 Move instance configuration into separate subpackage
