+++ server/src/main/scala/de/frosner/broccoli/instances/conf/FileSystemInstanceStorageConfiguration.scala
@@ -1,4 +1,4 @@
-package de.frosner.broccoli.instances
+package de.frosner.broccoli.instances.conf
 
 import com.typesafe.config.Config
