+++ server/src/main/scala/de/frosner/broccoli/services/InstanceService.scala
@@ -7,6 +7,7 @@ import javax.inject.{Inject, Singleton}
 import cats.data.EitherT
 import cats.instances.future.{getClass, _}
 import de.frosner.broccoli.instances._
+import de.frosner.broccoli.instances.conf.InstanceConfiguration
 import de.frosner.broccoli.models.JobStatus.JobStatus
 import de.frosner.broccoli.models._
 import de.frosner.broccoli.nomad.NomadClient
